feat: re-run style generation with current config
Seems that the automatic style generation didn't run for a while, but changes were done
in between. The changes which would have been done automatically by the plugin were done
manually and in some case changes in the configs were missed:

The SLDs for HB_Habitat were removed, but the respective style and layer were not deleted in
the config.style file (this PR https://github.com/wetransform-os/portrayal-repository/pull/23).
The automatic generation process failed as HB_Habitat was still configured but no SLD exists.
This is adapted here and the config of HB_Habitat layer and style is removed.

Additionally the GE_Borehole was added manually with this PR
https://github.com/wetransform-os/portrayal-repository/pull/29
But it was missed to add an entry in the generated.style file.

The changes related to 'tn-ro_4.0_1:RoadNode' -> 'tn-ro_4.0:RoadNode' in the config.xml file
seems to be related to this PR https://github.com/wetransform-os/portrayal-repository/pull/28
It was missed to do the respective change in the mentioned PR.

US_EnvironmentalManagementFacilitySite was removed and correct is US_EnvironmentalManagementFacility.
The change was done in this PR https://github.com/wetransform-os/portrayal-repository/pull/25
Seems that it was just missed to rename the feature type manually in the config.xmi file.

The tag 'production' for BU_Building and BU_BuildingPart was added with this PR
https://github.com/wetransform-os/portrayal-repository/pull/22
But it was missed to add the tag manually in the generated.style file.

SVC-280